se agrega cambios

+++ ApiTaller.Domain/Dtos/Product/GetProductDto.cs
@@ -9,8 +9,8 @@ namespace ApiTaller.Domain.Dtos.Product
     {
         public int Id { get; set; }
         public string ProductName { get; set; }
-        public int Price { get; set; }
-        public int SalePrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal SalePrice { get; set; }
         public string Code { get; set; }
         public string Reference { get; set; }
         public string Description { get; set; }
